requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-localflavor==4.0
django-photologue==3.17
django-recurrence==1.11.1
django-sortedm2m==3.1.1
ExifRead==3.0.0
fontawesomefree==6.4.2
idna==3.4
npm==0.1.1
optional-django==0.1.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.6
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==7.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
